+++ ta4j-core/src/test/java/org/ta4j/core/indicators/averages/SGMAIndicatorTest.java
@@ -23,6 +23,7 @@
  */
 package org.ta4j.core.indicators.averages;
 
+import static org.junit.Assert.*;
 import static org.ta4j.core.TestUtils.*;
 
 import org.junit.Test;
@@ -35,29 +36,48 @@ import org.ta4j.core.mocks.MockIndicator;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-public class SMMAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
+public class SGMAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
 
-    public SMMAIndicatorTest(NumFactory numFactory) {
+    public SGMAIndicatorTest(NumFactory numFactory) {
         super(numFactory);
     }
 
     @Test
-    public void smmaIndicatorTest() {
+    public void sgmaIndicatorTest() {
 
-        MockIndicator mock = CsvTestUtils.getCsvFile(SMMAIndicatorTest.class, "SMMA.csv", numFactory);
+        MockIndicator mock = CsvTestUtils.getCsvFile(SGMAIndicatorTest.class, "SGMA.csv", numFactory);
 
         BarSeries barSeries = mock.getBarSeries();
 
-        SMMAIndicator ma = new SMMAIndicator(new ClosePriceIndicator(barSeries), 10);
+        SGMAIndicator ma = new SGMAIndicator(new ClosePriceIndicator(barSeries), 9, 2);
 
         for (int i = 0; i < barSeries.getBarCount(); i++) {
 
             Num expected = mock.getValue(i);
             Num value = ma.getValue(i);
-
             assertNumEquals(expected.doubleValue(), value);
-
         }
     }
 
+    @Test(expected = IllegalArgumentException.class)
+    public void evenBarCountThrowsException() {
+        MockIndicator mock = CsvTestUtils.getCsvFile(SGMAIndicatorTest.class, "SGMA.csv", numFactory);
+
+        BarSeries barSeries = mock.getBarSeries();
+
+        SGMAIndicator ma = new SGMAIndicator(new ClosePriceIndicator(barSeries), 10, 2);
+
+        fail("Should have thrown an exception");
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void barCountShouldBeGreaterThanPolynomialOrderThrowsException() {
+        MockIndicator mock = CsvTestUtils.getCsvFile(SGMAIndicatorTest.class, "SGMA.csv", numFactory);
+
+        BarSeries barSeries = mock.getBarSeries();
+
+        SGMAIndicator ma = new SGMAIndicator(new ClosePriceIndicator(barSeries), 3, 5);
+
+        fail("Should have thrown an exception");
+    }
 }